requirements.txt
annotated-types==0.7.0
bitarray==2.9.2
blinker==1.8.2
cbor2==5.6.5
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
cytoolz==0.12.3
Deprecated==1.2.14
ecdsa==0.19.0
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
Flask==3.0.3
Flask-Limiter==3.8.0
hexbytes==1.2.1
idna==3.8
importlib_resources==6.4.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ntplib==0.4.0
ordered-set==4.1.0
packaging==24.1
parsimonious==0.10.0
pluggy==1.5.0
ply==3.11
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rlp==4.0.1
six==1.16.0
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.4
wrapt==1.16.0
ruff>=0.3
mypy>=1.8
